final da estilização
